+++ app/utils/server/index.ts
@@ -1,6 +1,8 @@
 import type { CronOptions } from 'croner';
 import { Cron } from 'croner';
 import type { H3Event } from 'h3';
+import { dirname } from 'node:path';
+import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'node:fs';
 
 export function defineCronJob(pattern: string, func: () => any, options?: CronOptions & { runOnInit?: true }): Promise<Cron>;
 export function defineCronJob(pattern: string, func: () => any, options: CronOptions & { runOnInit: false }): Cron;
@@ -45,6 +47,16 @@ export function getVATSIMIdentHeaders(): Record<string, string> {
     return {
         'User-Agent': token,
     };
+}
+
+export function writeJsonFile(path: string, data: unknown) {
+    mkdirSync(dirname(path), { recursive: true });
+    writeFileSync(path, JSON.stringify(data));
+}
+
+export function readJsonFile<T>(path: string): T | null {
+    if (!existsSync(path)) return null;
+    return JSON.parse(readFileSync(path, 'utf8')) as T;
 }
 
 export function getRedirectURL(event: H3Event) {
