+++ docs/ai/decisions.md
@@ -48,4 +48,4 @@
 - Kafka worker health logging is aggregated in `app/utils/server/worker/kafka.ts` on a 5s window. Keep it windowed instead of per-message so high-volume VATSIM streams do not flood stdout. Use Kafka message timestamp age to detect real-time delay, processing duration to detect slow handlers, dropped stale messages to detect the existing 10s cutoff, and broker latest offsets for partition lag.
 - ATC duplicating settings in `app/utils/server/vatsim/atc-duplicating.ts` use the V2 data shape: `description`, `authorCid`, `prefixes`, `suffixes`, optional `matchRegex`, and `mapping`. Prefer `prefixes`/`suffixes` for simple callsign families; keep `matchRegex` only for rules with numeric or otherwise constrained middle segments.
 - Kubernetes MySQL deployments intentionally pin the server image to the current 9.7 line and pass runtime memory options through container args, so existing PVC data is reused while `mysqld` starts with a 128M InnoDB buffer pool, 80 max connections, and disabled `performance_schema`. DB `resources` are left unset because the cluster uses fixed node placement and the explicit MySQL runtime settings are the memory-control mechanism.
-- Navigraph worker keeps only short navdata in memory after initialization. Full navdata is materialized into versioned JSON files under `app/data/navigraph-cache/<type>-<cycle>` and served from there, with a `ready` marker preventing reuse of partial caches and old cycle folders removed after a successful new cache build.
+- Navigraph worker keeps only short navdata in memory after initialization. Full navdata is materialized into versioned JSON files under `app/data/navigraph-cache/<type>-<cycle>` and served from there, with a `ready` marker preventing reuse of partial caches and old cycle folders removed after a successful new cache build. Navigraph-specific cache pathing/materialization helpers live in `app/utils/server/navigraph/cache.ts`; generic JSON file helpers live in `app/utils/server/index.ts`.
